logger.baseData: return an allocated map when 'extra' is positive
Previosly, baseData returned nil if the logger's data was nil and no
data was passed as an argument to the log method. This meant that the
Error and Fatal log methods had to handle the case where the returned
data was nil.  Respecting the 'extra' argument to baseData simplifies
the code with an immeasurable impact on performance.